+++ src/main/kotlin/com/dphong/likescash/api/product/model/RegisterProductRequest.kt
@@ -1,7 +1,7 @@
 package com.dphong.likescash.api.product.model
 
-import com.dphong.likescash.domain.Member
-import com.dphong.likescash.domain.Product
+import com.dphong.likescash.domain.member.Member
+import com.dphong.likescash.domain.product.Product
 
 data class RegisterProductRequest(val name: String) {
