+++ src/main/kotlin/com/dphong/likescash/api/product/service/RegisterProductService.kt
@@ -4,7 +4,7 @@ import com.dphong.likescash.api.product.model.RegisterProductRequest
 import com.dphong.likescash.api.product.model.RegisterProductResponse
 import com.dphong.likescash.api.product.model.RegisterProductStatus
 import com.dphong.likescash.common.reponse.StatusDataResult
-import com.dphong.likescash.domain.MemberRole
+import com.dphong.likescash.domain.member.MemberRole
 import com.dphong.likescash.repository.member.MemberRepository
 import com.dphong.likescash.repository.product.ProductRepository
 import org.springframework.stereotype.Service
