+++ src/main/kotlin/com/dphong/likescash/domain/member/MemberRole.kt
@@ -1,4 +1,4 @@
-package com.dphong.likescash.domain
+package com.dphong.likescash.domain.member
 
 enum class MemberRole {
     USER, ADMIN, SELLER
