+++ src/main/kotlin/com/dphong/likescash/repository/product/ProductJpaRepository.kt
@@ -1,6 +1,6 @@
 package com.dphong.likescash.repository.product
 
-import com.dphong.likescash.domain.Product
+import com.dphong.likescash.domain.product.Product
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface ProductJpaRepository : JpaRepository<Product, Long>
