+++ src/test/kotlin/com/dphong/likescash/api/product/controller/ProductControllerTest.kt
@@ -4,7 +4,7 @@ import com.dphong.likescash.BaseWebMvcTest
 import com.dphong.likescash.api.product.model.ProductDetails
 import com.dphong.likescash.api.product.service.ProductGetter
 import com.dphong.likescash.common.reponse.DataResult
-import com.dphong.likescash.domain.Product
+import com.dphong.likescash.domain.product.Product
 import com.dphong.likescash.fixtures.MemberFixture
 import org.junit.jupiter.api.Test
 import org.mockito.BDDMockito.given
