+++ src/test/kotlin/com/dphong/likescash/fixtures/MemberFixture.kt
@@ -1,7 +1,7 @@
 package com.dphong.likescash.fixtures
 
-import com.dphong.likescash.domain.Member
-import com.dphong.likescash.domain.MemberRole
+import com.dphong.likescash.domain.member.Member
+import com.dphong.likescash.domain.member.MemberRole
 
 object MemberFixture {
